requirements.txt
asgiref==3.5.0
Django==4.0.2
django-cleanup==6.0.0
django-cors-headers==3.11.0
djangorestframework==3.13.1
Pillow==9.0.1
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
